tests are moved into an integrationTests folder
