+++ tasks/webpack-build-task/tests/integrationTests/mockRunnerDefinitions/shared/taskMockRunnerFactory.ts
@@ -2,7 +2,7 @@ import { TaskMockRunner } from "vsts-task-lib/mock-run";
 import * as path from "path";
 
 const createTaskMockRunner = () => {
-    return new TaskMockRunner(path.join(__dirname, "../../../index.js"));
+    return new TaskMockRunner(path.join(__dirname, "../../../../index.js"));
 };
 
 export default createTaskMockRunner;
