+++ tasks/webpack-build-task/tests/integrationTests/mockRunnerDefinitions/shouldCompileSimpleWebpackProject.ts
@@ -2,6 +2,6 @@ import runTestTask from "./shared/testTaskRunner";
 import * as path from "path";
 
 runTestTask({
-    workingFolder: path.resolve(__dirname, "../../../../samples/webpack-3"),
+    workingFolder: path.resolve(__dirname, "../../../../../samples/multiple-webpack-build-steps/webpack-project-one"),
     mockWriteFile: true
 });
