+++ tasks/webpack-build-task/tests/integrationTests/mockRunnerDefinitions/shouldCompileWebpack2Project.ts
@@ -2,6 +2,6 @@ import runTestTask from "./shared/testTaskRunner";
 import * as path from "path";
 
 runTestTask({
-    workingFolder: path.resolve(__dirname, "../../../../samples/webpack-ts-config"),
+    workingFolder: path.resolve(__dirname, "../../../../../samples/webpack-2"),
     mockWriteFile: true
 });
