+++ tasks/webpack-build-task/tests/integrationTests/mockRunnerDefinitions/shouldCompileWebpack4Project.ts
@@ -2,6 +2,6 @@ import runTestTask from "./shared/testTaskRunner";
 import * as path from "path";
 
 runTestTask({
-    workingFolder: path.resolve(__dirname, "../../../../samples/webpack-3-with-issues"),
+    workingFolder: path.resolve(__dirname, "../../../../../samples/webpack-4"),
     mockWriteFile: true
 });
