+++ tasks/webpack-build-task/tests/integrationTests/mockRunnerDefinitions/shouldCompileWebpack4ProjectWithDisplayNone.ts
@@ -2,7 +2,7 @@ import runTestTask from "./shared/testTaskRunner";
 import * as path from "path";
 
 runTestTask({
-    workingFolder: path.resolve(__dirname, "../../../../samples/webpack-3"),
+    workingFolder: path.resolve(__dirname, "../../../../../samples/webpack-4"),
     mockWriteFile: true,
     webpackCliArguments: "--display \"none\""
 });
