+++ tasks/webpack-build-task/tests/integrationTests/mockRunnerDefinitions/shouldCompileWebpackTypeScriptConfigProject.ts
@@ -2,6 +2,6 @@ import runTestTask from "./shared/testTaskRunner";
 import * as path from "path";
 
 runTestTask({
-    workingFolder: path.resolve(__dirname, "../../../../samples/webpack-4"),
+    workingFolder: path.resolve(__dirname, "../../../../../samples/webpack-ts-config"),
     mockWriteFile: true
 });
