+++ tasks/webpack-build-task/tests/integrationTests/mockRunnerDefinitions/shouldHideErrorsInCaseOfNoneDisplayOption.ts
