+++ tasks/webpack-build-task/tests/integrationTests/mockRunnerDefinitions/shouldHideWarningsInCaseOfNoneDisplayOption.ts
