+++ tasks/webpack-build-task/tests/integrationTests/mockRunnerDefinitions/shouldReplaceBracketsToParenthesisInTaskName.ts
