+++ tasks/webpack-build-task/tests/integrationTests/mockRunnerDefinitions/shouldSucceedIfThereAreErrorsButTreatedAsInfo.ts
