+++ tasks/webpack-build-task/tests/integrationTests/mockRunnerDefinitions/shouldSucceedIfThereAreWarningsButTreatedAsInfo.ts
