+++ tasks/webpack-build-task/tests/integrationTests/shouldCompileSimpleWebpackProject.ts
