+++ tasks/webpack-build-task/tests/integrationTests/shouldCompileSimpleWebpackProjectWithIssues.ts
