+++ tasks/webpack-build-task/tests/integrationTests/shouldCompileWebpack2Project.ts
