+++ tasks/webpack-build-task/tests/integrationTests/shouldCompileWebpack3Project.ts
