+++ tasks/webpack-build-task/tests/integrationTests/shouldCompileWebpack3ProjectWithIssues.ts
