+++ tasks/webpack-build-task/tests/integrationTests/shouldCompileWebpack4Project.ts
