+++ tasks/webpack-build-task/tests/integrationTests/shouldCompileWebpackTypeScriptConfigProject.ts
