+++ tasks/webpack-build-task/tests/integrationTests/shouldFailIfThereAreErrorsAndWarnings.ts
