+++ tasks/webpack-build-task/tests/integrationTests/shouldFailIfThereAreWarningsButTreatedAsErrors.ts
