+++ tasks/webpack-build-task/tests/integrationTests/shouldPartiallySucceedIfThereAreErrorsButTreatedAsWarning.ts
