+++ tasks/webpack-build-task/tests/integrationTests/shouldReplaceBracketsToParenthesisInTaskName.ts
