+++ tasks/webpack-build-task/tests/integrationTests/shouldReportErrorDetailInCaseOfWebpackBuildFailure.ts
