+++ tasks/webpack-build-task/tests/integrationTests/shouldSucceedIfNoDisplayNameDefined.ts
