+++ tasks/webpack-build-task/tests/integrationTests/shouldSucceedIfThereAreErrorsButTreatedAsInfo.ts
