+++ tasks/webpack-build-task/tests/integrationTests/shouldSucceedIfThereAreWarningsButTreatedAsInfo.ts
